requirements.txt
flask
pandas
requests
python-dateutil
python-dotenv